Style: created 2 tabs component. Edited Followers/Following block
